requirements.txt
imageio==2.9.0
lpips==0.1.3
matplotlib==3.4.1
numpy==1.19.0
Pillow==8.1.1
PyYAML==6.0.1
scheduler==0.8.8
seaborn==0.13.2
sft==0.1.1
thop==0.0.31.post2005241907
torch==1.8.0
torch_summary==1.4.5
torchvision==0.9.0
tqdm==4.60.0
utils2==0.1.2
~umpy==1.24.4
